+++ src/main/java/cn/huohuas001/huHoBot/NetEvent/AddAllowList.java
@@ -8,7 +8,7 @@ public class AddAllowList extends EventRunner {
     public boolean run() {
         String XboxId = body.getString("xboxid");
         FileConfiguration config = this.getConfig();
-        String command = config.getString("addWhiteListCmd").replace("{name}", XboxId);
+        String command = config.getString("whiteList.add").replace("{name}", XboxId);
         runCommand(command);
         String name = HuHoBot.getPlugin().getServerName();
         respone(name + "已接受添加名为" + XboxId + "的白名单请求", "success");