+++ src/main/java/cn/huohuas001/huHoBot/NetEvent/DelAllowList.java
@@ -8,7 +8,7 @@ public class DelAllowList extends EventRunner {
     public boolean run() {
         String XboxId = body.getString("xboxid");
         FileConfiguration config = this.getConfig();
-        String command = config.getString("delWhiteListCmd").replace("{name}", XboxId);
+        String command = config.getString("whiteList.del").replace("{name}", XboxId);
         runCommand(command);
         String name = HuHoBot.getPlugin().getServerName();
         respone(name + "已接受删除名为" + XboxId + "的白名单请求", "success");